requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# AI / LLM Orchestration
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
groq>=0.9.0

# Vector Database
qdrant-client>=1.9.0
fastembed>=0.3.0

# YouTube Transcript
youtube-transcript-api>=0.6.2

# Data Handling
pandas>=2.2.0
pydantic>=2.6.0
python-dotenv>=1.0.1

# HTTP client (used by Streamlit → FastAPI calls)
httpx>=0.27.0
requests>=2.31.0

# UI
streamlit>=1.35.0

# Utilities
tenacity>=8.2.0
